project/requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1
google-generativeai==0.3.2
textblob==0.18.0
pandas==2.2.1
matplotlib==3.8.3
wordcloud==1.9.3
nltk==3.8.1
gensim==4.3.2
scikit-learn==1.4.1
PyPDF2==3.0.1
docx2txt==0.8